+++ test/test_mccontext.jl
@@ -8,10 +8,19 @@
 
     @test ctx.rng == Random.Xoshiro(123)
 
+    register_observable!(ctx, :test2; binsize = 2)
+
     Carlo.measure!(ctx, :test, 2.0)
+
+    Carlo.measure!(ctx, :test2, 2.0)
+    Carlo.measure!(ctx, :test2, 3)
+    Carlo.measure!(ctx, :test2, 4)
 
     tmp_hdf5_file() do file
         Carlo.write_measurements!(ctx, open_group(file, "/"))
+
+        @test read(file, "/observables/test2/samples") ≈ [2.5]
+        @test read(file, "/observables/test2/bin_length") ≈ 2
     end
 
     @test test_checkpointing(ctx, type = Carlo.MCContext{Random.Xoshiro})